refactor(mcp): decompose god module into 8 handler submodules
Move all 35 MCP tool implementations from the monolithic mod.rs (1310 lines)
into 8 category-specific handler modules under handlers/:

- assets.rs: download_assets (1 tool)
- obsidian.rs: detect_obsidian_vault, build_obsidian_uri, open_in_obsidian (3)
- url_utils.rs: validate_url, extract_domain, normalize_url, match_url_pattern, is_internal_link, url_to_file_path (6)
- security.rs: detect_waf, verify_waf_integrity, list_waf_providers, get_scrape_metrics (4)
- content.rs: clean_html, convert_html_to_markdown, extract_links, highlight_code_blocks, convert_wiki_links, generate_frontmatter, generate_rich_metadata (7)
- export.rs: export_file, export_jsonl, export_vector, process_export_pipeline (4)
- scraping.rs: scrape_url, scrape_with_options, scrape_batch, crawl_site, crawl_with_sitemap, discover_urls, discover_sitemap, detect_spa (8)
- ai.rs: semantic_cleaner, search_obsidian (2, feature-gated bodies)

Each handler uses #[tool_router(router = tool_router_NAME, vis = "pub")] to
generate a category-specific router. build_tool_router() combines all 8 via +.

mod.rs reduced from 1310 → 83 lines. ServerHandler trait implemented manually
using self.tool_router (combined field) instead of #[tool_handler] macro, since
the macro's default Self::tool_router() would only see the (now empty) local
router.
